Support 'additive' imports where the existing records are not deleted.

+++ src/Importer.php
@@ -179,7 +179,9 @@ class Importer {
      */
     public function run() {
 
-        $this->delete();
+        if (!$this->settings('additive imports')) {
+          $this->delete();
+        }
         $this->insert();
     }
